Move operators to separate files
