+++ blender_source/MH_Community/operators/posesync.py
@@ -1,0 +1,33 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
+import bpy
+
+class PoseSyncOperator(bpy.types.Operator):
+    """Synchronize the pose of the skeleton of a human with MH.  Requirements:\n\nMust be the Default armature.\nMust be exported in decimeters to allow location translation."""
+    bl_idname = "mh_community.sync_pose"
+    bl_label = "Synchronize MH Pose"
+    bl_options = {'REGISTER', 'UNDO'}
+
+    def execute(self, context):
+        from ..rig_info import RigInfo
+        from ..mh_sync.sync_pose import SyncPose
+
+        armature = context.object
+        rigInfo = RigInfo.determineRig(armature)
+        if rigInfo.determineExportedUnits != 'DECIMETERS' and not bpy.context.scene.MhNoLocation:
+            self.report({'ERROR'}, 'Location translation only possible when exported in decimeters to match MakeHuman.')
+            return {'FINISHED'}
+
+        SyncPose()
+        return {'FINISHED'}
+    # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+    @classmethod
+    def poll(cls, context):
+        from ..rig_info import RigInfo
+        ob = context.object
+        if ob is None or ob.type != 'ARMATURE': return False
+
+        # can now assume ob is an armature
+        rigInfo = RigInfo.determineRig(ob)
+        return rigInfo is not None and rigInfo.isPoseCapable()
